+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### [Unreleased][HEAD]
 
-   * **portal**: add `commitItem()` function for item file upload commit
+   * **portal**: add `commitItemUpload()` and `cancelItemUpload()` functions for item file upload
    * **portal**: add `overwrite` paramter for item creation
 
 ## [2.4.0] - August 29th 2019